+++ src/main.rs
@@ -8,7 +8,9 @@ extern crate serde;
 extern crate serde_derive;
 extern crate serde_json;
 extern crate tar;
+extern crate textwrap;
 extern crate toml;
+extern crate which;
 extern crate zip;
 
 use clap::{App, SubCommand};
@@ -24,6 +26,7 @@ mod utils {
     pub mod http;
     pub mod nodejs;
     pub mod process;
+    pub mod ssh;
 }
 
 fn main() {